Add basic structure of eagle-hadoop-health module
