+++ eagle-hadoop-health/src/main/java/org/apache/eagle/health/jobs/HBaseHealthCheckJob.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.eagle.health.job;
+package org.apache.eagle.health.jobs;
 
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.ListMultimap;
