+++ eagle-hadoop-health/src/main/java/org/apache/eagle/health/jobs/HDFSHealthCheckJob.java
@@ -1,0 +1,51 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.eagle.health.jobs;
+
+import org.apache.eagle.app.job.MonitorJob;
+import org.apache.eagle.app.job.MonitorResult;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.fs.FileSystem;
+import org.quartz.JobExecutionContext;
+import org.quartz.JobExecutionException;
+
+import java.io.IOException;
+
+/**
+ * hdfs haadmin -checkHealth <serviceId>
+ */
+public class HDFSHealthCheckJob extends MonitorJob {
+
+    @Override
+    protected void prepare(JobExecutionContext context) throws JobExecutionException {
+        try {
+            FileSystem fileSystem = FileSystem.get(new Configuration());
+        } catch (IOException e) {
+            throw new JobExecutionException(e);
+        }
+    }
+
+    @Override
+    protected MonitorResult execute() throws JobExecutionException, Exception {
+        return null;
+    }
+
+    @Override
+    protected void close() throws JobExecutionException {
+
+    }
+}
