+++ eagle-hadoop-health/src/main/java/org/apache/eagle/health/jobs/HDFSMissingBlockCheckJob.java
@@ -15,27 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.eagle.health.job;
+package org.apache.eagle.health.jobs;
 
 import org.apache.eagle.app.job.MonitorJob;
 import org.apache.eagle.app.job.MonitorResult;
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
 
-/**
- * Do check for:
- * 1) Resource Manager:
- * 2) Node Manager: http://${NODE_MANAGER_HOST}:8042/ws/v1/node/info.
- */
-public class YARNHealthCheckJob extends MonitorJob {
+public class HDFSMissingBlockCheckJob extends MonitorJob {
     @Override
     protected MonitorResult execute() throws JobExecutionException, Exception {
+        // Compare current missing block and previous version of under-replicated
         return null;
     }
 
     @Override
     protected void prepare(JobExecutionContext context) throws JobExecutionException {
-
+        // Previous dump under-replication blocks
     }
 
     @Override
