+++ eagle-hadoop-health/src/main/java/org/apache/eagle/health/jobs/YARNHealthCheckJob.java
@@ -5,45 +5,41 @@
  * The ASF licenses this file to You under the Apache License, Version 2.0
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
- * <p>
+ * <p/>
  * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
+ * <p/>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.eagle.health.job;
+
+package org.apache.eagle.health.jobs;
+
+import org.apache.eagle.app.job.MonitorJob;
+import org.apache.eagle.app.job.MonitorResult;
+import org.quartz.JobExecutionContext;
+import org.quartz.JobExecutionException;
 
 /**
- * Healthy Checker Event/Status Type.
+ * Do check for:
+ * 1) Resource Manager:
+ * 2) Node Manager: http://${NODE_MANAGER_HOST}:8042/ws/v1/node/info.
  */
-public enum EventType {
-    OK,
+public class YARNHealthCheckJob extends MonitorJob {
+    @Override
+    protected MonitorResult execute() throws JobExecutionException, Exception {
+        return null;
+    }
 
-    UNKNOWN_ERROR,
-    TIMEOUT_ERROR,
-    EXECUTION_ERROR,
-    INTERRUPTED_ERROR,
+    @Override
+    protected void prepare(JobExecutionContext context) throws JobExecutionException {
 
-    MAPRED_JOB_ERROR,
+    }
 
-    HDFS_CONNECT_ERROR,
-    HDFS_READ_ERROR,
-    HDFS_WRITE_ERROR,
-    HDFS_DELETE_ERROR,
-    HDFS_CREATE_ERROR,
+    @Override
+    protected void close() throws JobExecutionException {
 
-    HBASE_MASTER_NOT_RUNNING,
-    HBASE_READ_TABLE_ERROR,
-    HBASE_CREATE_TABLE_ERROR,
-    HBASE_CREATE_REGION_ERROR,
-    HBASE_MOVE_REGION_ERROR,
-    HBASE_WRITE_REGION_ERROR,
-    HBASE_READ_REGION_ERROR,
-
-    ZOOKEEPER_ERROR,
-    ZOOKEEPER_CONNECTION_ERROR,
-    ZOOKEEPER_IO_ERROR,
+    }
 }
